Renamed variables, added CRON_ENABLED

+++ entrypoint.sh
@@ -8,20 +8,21 @@ cp /usr/share/zoneinfo/$TZ /etc/localtime
 echo $TZ > /etc/timezone
 
 # Configure rclone
-if [[ $( rclone $RCLONE_OPTS config show | grep "empty config" | wc -l ) != "0" ]]; then
+if [[ $( rclone $CONFIG_OPTS config show | grep "empty config" | wc -l ) != "0" ]]; then
   echo "$( date ) Configuring rclone"
-  rclone $RCLONE_OPTS config
+  rclone $CONFIG_OPTS config
 fi
 
-# Tidy up previous sync jobs
-rm -f /tmp/sync.pid
+if [[ "$CRON_ENABLED" != "1" ]; then
+  /rclone.sh
+else
+  # Setup cron schedule
+  crontab -d
+  echo "$CRON /rclone.sh >> /dev/stdout 2>&1" > /tmp/crontab.tmp
+  crontab /tmp/crontab.tmp
+  rm /tmp/crontab.tmp
 
-# Setup cron schedule
-crontab -d
-echo "$CRON /sync.sh >> /dev/stdout 2>&1" > /tmp/crontab.tmp
-crontab /tmp/crontab.tmp
-rm /tmp/crontab.tmp
-
-# Start cron
-echo "$( date ) Starting cron"
-crond -b -l 0 -L /dev/stdout
+  # Start cron
+  echo "$( date ) Starting cron"
+  crond -b -l 0 -L /dev/stdout
+fi
